Most metrics seem reasonable now.
Our junction end coords are now 1 less than what tophat / spanki report.  This is because this is the inclusive position (i.e. last position in the junction), rather than the position of the start of the downstream anchor.  Not sure what to do about this yet.

Some of the anchor metrics are way out from what one might expect... this is because they describe the full alignments and are not bounded by up/down stream junctions within the alignment.  Again, not sure whether to bound them or not at the moment.