Updated to 0.12.0
Reworking front end to remove settings file from prepare stage and dependencies on this in later stages.

Full pipeline now works properly, with BAM filtering and optional stage.

Prepare now checks to see if any sequence in the genome is longer than the maximum for BAI indexing.  If using BAI indexing and such a sequence is detected then an error is thrown.