This commit will break existing pipelines for the filtering stage at present.
The changes allow for ML based filtering using a pre-built model.  The catch is that we have to use python to do this, so we call out to python code by default to do the filtering.  I haven't written the python filtering script yet!  (Although there is a script that can build models).  We retain the ability to use rule based filtering, but the user will explicitly have to specify a rules file to the filtering tool to do this.

Had to modify the configuration in order to make all the changes work.