Updating the junction analysis tool to include reliable 2 raw alignment ratio scores.
Also removed some of the features going into the prediction tool which seems to ensure that all features are now contributing to the decision making process under all the situations I've tested so far.  Performance seems on par with finesplice, but we are still a little way behind truesight and soapsplice with this version.