Ensuring the python configuration is more stable.
Now if the user specifies to install to a custom path, python packages get installed to that location, and the user will have to set a custom PYTHONPATH at runtime to access the packages.

If the user installs to a system directory (default), the python packages get installed to the currently accessible python3 instance's site packages (this may not be in /usr/local).

In order to ensure dependable behavior for embedded scripts we install into <prefix>/share/portcullis/scripts/portcullis/...